task-3/agent-team/pyproject.toml
[project]
name = "agent-team"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "python-fasthtml>=0.12.47",
]
